Fixed a traceback I keep seeing in the logs
When reloading the page, it was reusing the previous channel ID. That's
because the browser was caching the request to fetch the channel ID. That's
because I was using a GET to make the request. I switched to using POST, and
the traceback in the logs went away. I also verified that each page reload is
now getting a fresh channel ID. This bug was painful!